float lr

+++ Train/TPU_training.py
@@ -32,7 +32,7 @@ import os
 parser = argparse.ArgumentParser(description='Train the model using Colab TPUs.')
 
 parser.add_argument('--n_epochs', type=int, required=True, help='Number of epochs')
-parser.add_argument('--learning_rate', type=int, required=True, help='Learning rate')
+parser.add_argument('--learning_rate', type=float, required=True, help='Learning rate')
 parser.add_argument('--hr_shape', type=make_tuple, required=True, help='High resolution shape')
 parser.add_argument('--lr_shape', type=make_tuple, required=True, help='Low resolution shape')
 parser.add_argument('--train_hr_path', type=str, required=True, help='Path to training HR tfRecords in GCS')
